requirements.txt
aiohttp==3.9.1
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
build==1.0.3
CacheControl==0.13.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
cryptography==41.0.7
databases==0.8.0
distlib==0.3.8
dnspython==2.4.2
dulwich==0.21.7
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.108.0
fastapi-cache2==0.2.1
fastjsonschema==2.19.1
filelock==3.13.1
frozenlist==1.4.1
gevent==23.9.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jeepney==0.8.0
Jinja2==3.1.2
keyring==24.3.0
Mako==1.3.0
MarkupSafe==2.1.3
more-itertools==10.1.0
msgpack==1.0.7
multidict==6.0.4
orjson==3.9.10
packaging==23.2
pendulum==3.0.0
pexpect==4.9.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
ptyprocess==0.7.0
pycparser==2.21
pydantic==2.5.3
pydantic-extra-types==2.2.0
pydantic-settings==2.1.0
pydantic_core==2.14.6
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
rapidfuzz==3.6.0
redis==4.6.0
requests==2.31.0
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.50
starlette==0.32.0.post1
time-machine==2.13.0
tomli==2.0.1
tomlkit==0.12.3
trove-classifiers==2023.11.29
typing_extensions==4.9.0
tzdata==2023.3
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.25.0
uvloop==0.19.0
virtualenv==20.25.0
watchfiles==0.21.0
websocket-client==1.7.0
yarl==1.9.4
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
